Refactor: Move utility functions and JSON parser to dedicated files
- Deleted `subsonic_json_parser.h` and `subsonic_json_parser.cpp`, moving their contents to `utils/subsonic_json_parser.h` and `utils/subsonic_json_parser.cpp`.
- Deleted `time_utils.h` and `time_utils.cpp`, moving their contents to `utils/time_utils.h` and `utils/time_utils.cpp`.
- Deleted `utils.h` and `utils.cpp`, moving their contents to `utils/utils.h` and `utils/utils.cpp`.
- Deleted `metadata_utils.h` and `metadata_utils.cpp`, moving their contents to `utils/metadata_utils.h` and `utils/metadata_utils.cpp`.
- Deleted `string_util.h` and `string_util.cpp`, moving their contents to `utils/string_util.h` and `utils/string_util.cpp`.
- Updated include paths in `vfs.cpp` to reflect new locations of utility headers.
- Ensured all utility functions are properly namespaced under `subsonic`.
