+++ src/cache.cpp
@@ -3,7 +3,7 @@
 #include "cache.h"
 
 #include "config.h"
-#include "utils.h"
+#include "utils/utils.h"
 
 #include <SDK/configStore.h>
 #include <SDK/filesystem_helper.h>
