+++ src/config.cpp
@@ -2,7 +2,7 @@
 
 #include "config.h"
 
-#include "utils.h"
+#include "utils/utils.h"
 
 namespace {
