+++ src/foo_opensubsonic.vcxproj
@@ -274,12 +274,12 @@
     <ClInclude Include="playback.h" />
     <ClInclude Include="resource.h" />
     <ClInclude Include="stdafx.h" />
-    <ClInclude Include="string_util.h" />
-    <ClInclude Include="subsonic_json_parser.h" />
-    <ClInclude Include="metadata_utils.h" />
-    <ClInclude Include="time_utils.h" />
     <ClInclude Include="types.h" />
-    <ClInclude Include="utils.h" />
+    <ClInclude Include="utils\string_util.h" />
+    <ClInclude Include="utils\subsonic_json_parser.h" />
+    <ClInclude Include="utils\metadata_utils.h" />
+    <ClInclude Include="utils\time_utils.h" />
+    <ClInclude Include="utils\utils.h" />
     <ClInclude Include="vfs.h" />
   </ItemGroup>
   <ItemGroup>
@@ -302,11 +302,11 @@
       <PrecompiledHeader Condition="'$(Configuration)|$(Platform)'=='Release|ARM64EC'">Create</PrecompiledHeader>
       <PrecompiledHeader Condition="'$(Configuration)|$(Platform)'=='Release|ARM64'">Create</PrecompiledHeader>
     </ClCompile>
-    <ClCompile Include="string_util.cpp" />
-    <ClCompile Include="subsonic_json_parser.cpp" />
-    <ClCompile Include="metadata_utils.cpp" />
-    <ClCompile Include="time_utils.cpp" />
-    <ClCompile Include="utils.cpp" />
+    <ClCompile Include="utils\string_util.cpp" />
+    <ClCompile Include="utils\subsonic_json_parser.cpp" />
+    <ClCompile Include="utils\metadata_utils.cpp" />
+    <ClCompile Include="utils\time_utils.cpp" />
+    <ClCompile Include="utils\utils.cpp" />
     <ClCompile Include="vfs.cpp" />
   </ItemGroup>
   <ItemGroup>
