+++ src/http.cpp
@@ -2,7 +2,7 @@
 
 #include "http.h"
 
-#include "utils.h"
+#include "utils/utils.h"
 
 namespace {
