+++ src/utils/metadata_utils.cpp
@@ -1,4 +1,4 @@
-#include "stdafx.h"
+#include "../stdafx.h"
 
 #include "metadata_utils.h"
