+++ src/utils/string_util.h
