+++ src/utils/subsonic_json_parser.cpp
@@ -1,4 +1,4 @@
-#include "stdafx.h"
+#include "../stdafx.h"
 
 #include "subsonic_json_parser.h"
