+++ src/utils/utils.cpp
@@ -1,4 +1,4 @@
-#include "stdafx.h"
+#include "../stdafx.h"
 
 #include "utils.h"
 #include "string_util.h"
